+++ packages/ui-editor/package.json
@@ -30,6 +30,8 @@
     "styled-components": "^4.0.0"
   },
   "peerDependencies": {
+    "@edtr-io/core": "^0.9.3",
+    "@edtr-io/ui": "^0.9.3",
     "@fortawesome/fontawesome-svg-core": "^1.0.0",
     "@fortawesome/free-solid-svg-icons": "^5.0.0",
     "@fortawesome/react-fontawesome": "^0.1.0",
