+++ packages/ui-editor/src/components/primary-settings.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../theme'
+import { styled } from '@edtr-io/ui'
 
 export const PrimarySettings = styled.div({
   marginTop: '15px'
