+++ packages/ui-editor/src/index.ts
@@ -1,3 +1,2 @@
+export { styled } from '@edtr-io/ui'
 export * from './components'
-export * from './overlay'
-export * from './theme'
