+++ packages/ui-renderer/src/components/feedback.tsx
@@ -1,6 +1,5 @@
+import { styled } from '@edtr-io/ui'
 import * as React from 'react'
-
-import { styled } from '../theme'
 
 export class Feedback extends React.Component<FeedbackProps> {
   public render() {
