+++ packages/ui-renderer/src/index.ts
@@ -1,2 +1,2 @@
+export { styled } from '@edtr-io/ui'
 export * from './components'
-export * from './theme'
